refactor: split god-services into facade + sub-services (Items, Preorders, Shops, AI)
* fix(ci): delete Neon preview branch before recreate to fix 422 on synchronize

On PR force-push (synchronize), create-branch-action fails with 422
because the branch already exists. Delete first with continue-on-error
so fresh PRs (no existing branch) still work.


* refactor(preorders,items,shops,ai): split god-services into facade + sub-services

- Split 4 large service files into thin facade + domain sub-services.
  Facade preserves all public API signatures; no controller changes.
- preorders: CrudService, StatusService, FinancialService
- items: CrudService, SearchService, FacebookService, ExportService,
  VectorService
- shops: CrudService, AppearanceService, MembersService, AuditService
- ai: ChatService, CatalogService

* refactor(utils): extract shared utilities from service classes

- requireActiveShopId -> src/common/utils/require-active-shop.ts
- assertShopExists -> shops/utils/assert-shop-exists.ts
- mergeContactJson, mergeAppearanceJson, mergeLoyaltyJson ->
  shops/utils/merge-shop-json.ts
- hydrateAppearanceJson -> shops/utils/hydrate-appearance.ts
- clampCustomInstructions, parseJsonObject, mapProviderError ->
  ai/ai-utils.ts

Removes all cross-sub-service injection; dependency graph stays a DAG.

* fix(preorders,items,shops): address correctness issues found during refactor

- preorders: cancelled_at only written on CANCELLED transition; was
  incorrectly nulled on every non-CANCELLED update
- items: add shop_id + deleted_at:null to update/remove where clauses
  for defense-in-depth tenant isolation
- shops: declare explicit PrismaModule dependency in ShopsModule

* test(preorders,items,shops): add coverage for refactored services

- preorders: tenant scope enforcement (empty tenantId -> AUTH_FORBIDDEN),
  cancelled_at regression, PAID/REFUNDED with null member_id
- items: shop_id + deleted_at in closePreorder/reopenPreorder assertions
- common: requireActiveShopId unit spec (valid / undefined / empty)
- docs/ARCHITECTURE.md: document facade + sub-services pattern and
  sub-service isolation rule

---------
